# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kobai",
    "version": "0.0.0",
    "description": "kobai's HTTP surface. Two authenticated surfaces: `/admin`, behind a Merchant session, and `/store`, behind an API key (ADR-0020). `/health` is open, and is the only route that answers before migrations have applied.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "merchantSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Merchant session token, from `POST /admin/session`. Opens `/admin` and nothing else."
      },
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "kobai_pk_… | kobai_sk_…",
        "description": "An API key, from `POST /admin/api-keys`. Opens `/store` and nothing else. The prefix says which kind it is, with no lookup: `kobai_pk_` is publishable and safe in a browser, `kobai_sk_` is secret and is not."
      }
    },
    "schemas": {
...
      "Merchant": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MerchantIdentity"
          },
          {
            "type": "object",
            "properties": {
              "role": {
                "$ref": "#/components/schemas/RoleSummary"
              }
            },
            "required": [
              "role"
            ]
          }
        ]
      },
...
      "Refusal": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in prose."
          },
          "reason": {
            "type": "string",
            "description": "Machine-readable. Branch on this."
          }
        },
        "required": [
          "error",
          "reason"
        ]
      },
      "SessionRefusal": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "session-missing",
              "session-malformed",
              "session-unknown",
              "session-expired"
            ]
          }
        },
        "required": [
          "error",
          "reason"
        ]
      },
...
      "ServerError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CreateMerchantRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "A Role by name. Defaults to `owner`."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/merchants": {
      "post": {
        "summary": "Create a Merchant",
        "description": "Needs `merchant:write`, except on a deployment that holds no Merchant at all — the first Merchant claims it, anonymously and exactly once.",
        "security": [
          {},
          {
            "merchantSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMerchantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Merchant, and the Role they hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Merchant"
                }
              }
            }
          },
          "400": {
            "description": "The request does not fit this endpoint's schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "403": {
            "description": "The Merchant's Role does not hold the permission this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDenied"
                }
              }
            }
          },
          "409": {
            "description": "Somebody got there first: the address is taken, or the deployment is already claimed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/admin/session": {
      "post": {
        "summary": "Sign in",
        "description": "An unknown address and a wrong password are answered identically, and in the same time. Distinguishing them would turn this into a way to ask who works here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The session, with the token to present on every later request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedSession"
                }
              }
            }
          },
          "400": {
            "description": "The request does not fit this endpoint's schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "401": {
            "description": "Those credentials are not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentials"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Who the caller is",
        "description": "What the Admin asks first after a page load: who am I, and what may I do.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's Merchant and Role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Sign out",
        "description": "The row goes, so the token stops working on the very next request.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "204": {
            "description": "Signed out."
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/admin/store": {
      "get": {
        "summary": "Read the Store",
        "description": "One deployment is one Store (ADR-0005), so this takes no identifier.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "The Store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Store"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "403": {
            "description": "The Merchant's Role does not hold the permission this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDenied"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/products/{id}": {
      "get": {
        "summary": "Read a Product",
        "description": "One Product, with its Variants and each Variant's Prices.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "An identifier. Anything that is not one is not found."
            },
            "required": true,
            "description": "An identifier. Anything that is not one is not found.",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The Product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "403": {
            "description": "The Merchant's Role does not hold the permission this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDenied"
                }
              }
            }
          },
          "404": {
            "description": "No such Product exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}